requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
binance==0.3
binance-futures-connector==3.3.1
cachelib==0.9.0
ccxt==3.0.29
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
constantly==15.1.0
cryptography==39.0.2
dateparser==1.1.7
Flask==2.2.3
Flask-Caching==2.0.2
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.3.0
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==23.0.0
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywifi==1.1.12
redis==4.5.1
regex==2022.10.31
requests==2.28.2
service-identity==21.1.0
six==1.16.0
TA-Lib==0.4.25
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
zope.interface==6.0
